Fix scopes in policies and PEP8 style errors

+++ garage/tf/core/parameterized.py
@@ -83,9 +83,8 @@ class Parameterized(object):
                     self._cached_assign_placeholders[
                         param] = assign_placeholder
                 ops.append(self._cached_assign_ops[param])
-                feed_dict[
-                    self._cached_assign_placeholders[param]] = value.astype(
-                        dtype)
+                feed_dict[self._cached_assign_placeholders[
+                    param]] = value.astype(dtype)
                 if debug:
                     print("setting value of %s" % param.name)
             tf.get_default_session().run(ops, feed_dict=feed_dict)
